+++ src/main/java/com/github/hcsp/anno/Cache.java
@@ -5,5 +5,5 @@ import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Cache {
-    int cacheSeconds() default 1;
+    long cacheSeconds() default 1;
 }
